modular build function
